+++ wger/exercises/api/serializers.py
@@ -24,6 +24,8 @@ from django.db.models import Q
 
 # Third Party
 from actstream import action as actstream_action
+from easy_thumbnails.exceptions import InvalidImageFormatError
+from easy_thumbnails.files import get_thumbnailer
 from rest_framework import serializers
 
 # wger
@@ -120,6 +122,7 @@ class ExerciseImageSerializer(serializers.ModelSerializer):
 
     author_history = serializers.ListSerializer(child=serializers.CharField(), read_only=True)
     exercise_uuid = serializers.ReadOnlyField(source='exercise.uuid')
+    thumbnails = serializers.SerializerMethodField()
 
     class Meta:
         model = ExerciseImage
@@ -129,6 +132,7 @@ class ExerciseImageSerializer(serializers.ModelSerializer):
             'exercise',
             'exercise_uuid',
             'image',
+            'thumbnails',
             'is_main',
             'style',
             'license',
@@ -140,6 +144,23 @@ class ExerciseImageSerializer(serializers.ModelSerializer):
             'author_history',
             'is_ai_generated',
         )
+
+    def get_thumbnails(self, obj: ExerciseImage):
+        if not obj.image:
+            return None
+
+        request = self.context.get('request')
+        result = {}
+
+        thumbnailer = get_thumbnailer(obj.image)
+        try:
+            for alias in ('small', 'medium'):
+                opts = settings.THUMBNAIL_ALIASES[''][alias]
+                thumb = thumbnailer.get_thumbnail(opts)
+                result[alias] = request.build_absolute_uri(thumb.url)
+        except (InvalidImageFormatError, OSError):
+            return None
+        return result
 
 
 class ExerciseVideoSerializer(serializers.ModelSerializer):