+++ wger/exercises/tests/test_exercise_images.py
@@ -15,6 +15,7 @@
 # Django
 from django.core.files import File
 from django.test import SimpleTestCase
+from django.urls import reverse
 
 # wger
 from wger.core.tests import api_base_test
@@ -128,6 +129,21 @@ class MainImageTestCase(WgerTestCase):
         image.refresh_from_db()
 
         self.assertTrue(image.is_ai_generated)
+
+    def test_api_exposes_thumbnails(self):
+        """
+        Tests that the exercise image API exposes thumbnail URLs
+        """
+        translation = Translation.objects.get(pk=2)
+        pk = self.save_image(translation.exercise, 'protestschwein.jpg')
+
+        response = self.client.get(reverse('exerciseimage-detail', kwargs={'pk': pk}))
+        self.assertEqual(response.status_code, 200)
+
+        thumbnails = response.json()['thumbnails']
+        self.assertEqual(set(thumbnails.keys()), {'small', 'medium'})
+        for url in thumbnails.values():
+            self.assertTrue(url.startswith('http'))
 
     def test_replacing_image_keeps_new_file(self):
         """